projet/requirements.txt
flask-graphql
flask-migrate
flask-sqlalchemy
graphene
graphene-sqlalchemy
Flask-PyMongo=2.3.0
python-dotenv
graphene
flask-graphql
graphene-mongo